chore: update dev scripts - debounced watch - rm xpi before start - refactor start.mjs - add html and css - log miss fluent message for each locale

+++ scripts/start.mjs
@@ -24,8 +24,7 @@ if (!existsSync(profilePath)) {
 function prepareDevEnv() {
     const addonProxyFilePath = path.join(profilePath, `extensions/${addonID}`);
     const buildPath = path.resolve("build/addon");
-
-    function writeAddonProxyFile() {
+    if (!existsSync(addonProxyFilePath) || readFileSync(addonProxyFilePath, "utf-8") !== buildPath) {
         writeFileSync(addonProxyFilePath, buildPath);
         Logger.debug(
             `Addon proxy file has been updated. 
@@ -34,12 +33,9 @@ function prepareDevEnv() {
         );
     }
 
-    if (existsSync(addonProxyFilePath)) {
-        if (readFileSync(addonProxyFilePath, "utf-8") !== buildPath) {
-            writeAddonProxyFile();
-        }
-    } else {
-        writeAddonProxyFile();
+    const addonXpiFilePath = path.join(profilePath, `extensions/${addonID}.xpi`);
+    if (existsSync(addonXpiFilePath)) {
+        rmSync(addonXpiFilePath);
     }
 
     const prefsPath = path.join(profilePath, "prefs.js");
